requirements.txt
asgiref             3.4.1
certifi             2021.5.30
charset-normalizer  2.0.4
coreapi             2.3.3
coreschema          0.0.4
Django              3.2.6
django-rest-auth    0.9.5
djangorestframework 3.12.4
drf-yasg            1.20.0 
idna                3.2
inflection          0.5.1
itypes              1.2.0
Jinja2              3.0.1
MarkupSafe          2.0.1
openapi-codec       1.3.2
packaging           21.0
pip                 21.1.3
pyparsing           2.4.7
pytz                2021.1
requests            2.26.0
ruamel.yaml         0.17.16
ruamel.yaml.clib    0.2.6
setuptools          56.0.0
simplejson          3.17.5
six                 1.16.0
sqlparse            0.4.1
uritemplate         3.0.1
urllib3             1.26.6


python manage.py collectstatic 

You have requested to collect static files at the destination
location as specified in your settings:

    /Users/a1/Desktop/Arhituktura/BASE_DIR / "static"

This will overwrite existing files!
Are you sure you want to do this?